requirements.txt
pandas
selenium
webdriver-manager
pillow
pytesseract
opencv-python